+++ cpp/src/pegasus/dataset/identity.h
@@ -32,7 +32,7 @@ public:
   
   std::string dataset_path() const;
   std::string file_path() const;
-  std::vector<int64_t> col_ids() const;
+  std::vector<int> col_ids() const;
   int64_t row_group_id() const;
   int64_t num_rows() const;
   int64_t bytes() const;
@@ -62,7 +62,7 @@ public:
   std::string dataset_path_;
   std::string file_path_;
   int64_t partid;
-  std::vector<int64_t> col_ids_;
+  std::vector<int> col_ids_;
   int64_t row_group_id_;
   int64_t num_rows_;
   int64_t bytes_;
